Add corutine Collector and Artist fix 2
