+++ app/src/main/java/com/example/vinilos/data/repository/ArtistService.kt
@@ -1,5 +1,6 @@
 package com.example.vinilos.data.repository
 
+import com.example.vinilos.data.api.ApiService
 import com.example.vinilos.data.api.RetrofitBuilder
 import com.example.vinilos.data.model.ArtistResponse
 import kotlinx.coroutines.Dispatchers
@@ -10,7 +11,7 @@ class ArtistService {
 
     suspend fun getArtist():List<ArtistResponse> {
         return   withContext(Dispatchers.IO) {
-            val response = retrofit.create(ArtistApiClient::class.java).getAllArtist()
+            val response = retrofit.create(ApiService::class.java).getAllArtist()
             response.body() ?: emptyList()
         }
     }
