--- app/src/main/java/com/example/vinilos/data/repository/CollectorApiClient.kt
@@ -1,10 +1,0 @@
-package com.example.vinilos.data.repository
-
-import com.example.vinilos.data.model.CollectorResponse
-import retrofit2.Response
-import retrofit2.http.GET
-
-interface CollectorApiClient {
-    @GET("collectors")
-    suspend fun getAllCollectors(): Response<List<CollectorResponse>>
-}
