+++ app/src/main/java/com/example/vinilos/data/repository/CollectorService.kt
@@ -1,5 +1,6 @@
 package com.example.vinilos.data.repository
 
+import com.example.vinilos.data.api.ApiService
 import com.example.vinilos.data.api.RetrofitBuilder
 import com.example.vinilos.data.model.CollectorResponse
 import kotlinx.coroutines.Dispatchers
@@ -10,7 +11,7 @@ class CollectorService {
 
     suspend fun getCollectors():List<CollectorResponse> {
         return   withContext(Dispatchers.IO) {
-            val response = retrofit.create(CollectorApiClient::class.java).getAllCollectors()
+            val response = retrofit.create(ApiService::class.java).getAllCollectors()
             response.body() ?: emptyList()
         }
     }
