+++ app/src/main/java/com/example/vinilos/ui/main/viewmodel/CollectorViewModel.kt
@@ -3,7 +3,6 @@ package com.example.vinilos.ui.main.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.liveData
-import com.example.vinilos.data.model.Album
 import com.example.vinilos.data.repository.CollectorRepository
 import com.example.vinilos.utils.Resource
 import kotlinx.coroutines.Dispatchers
